+++ example/recent_playlists_example.dart
@@ -1,4 +1,3 @@
-import 'package:drift/drift.dart';
 import 'package:rekorddart/logger.dart';
 import 'package:rekorddart/rekorddart.dart';