+++ lib/rekorddart.dart
@@ -1,9 +1,11 @@
 /// A Dart library for accessing Rekordbox databases.
 library;
 
+export 'package:drift/drift.dart';
+
 export 'src/config.dart';
 export 'src/database/rekordbox_database.dart';
+export 'src/extensions/djmd_content_extensions.dart';
 export 'src/extensions/object_value_extensions.dart';
 export 'src/extensions/query_extensions.dart';
-export 'src/extensions/string_extensions.dart';
 export 'src/functions.dart';