+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- chore: export `package:drift` from this package, circumventing the need to add it as a dependency in your project.
+- docs: minor example updates, add outer join in `recent_tracks_example`.
+
 ## 1.0.0
 
 - Initial version.